// frontend/src/api/generated/model/commentResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CommentResponse {
  id: number;
  postId: number;
  authorId: number;
  authorUsername: string;
  parentId?: number;
  content: string;
  score: number;
  depth: number;
  deleted: boolean;
  createdAt: string;
  children: CommentResponse[];
}